FE: Setting up local storage
