+++ client/store/index.js
@@ -8,7 +8,9 @@ import singleProductReducer from './singleProduct'
 import cartReducer from './cart'
 import allUsersReducer from './allUsers'
 import localCartReducer from './localCart'
+import {loadState, saveState} from './localStorage'
 
+const localCart = loadState()
 const reducer = combineReducers({
   user,
   allProductsReducer,
@@ -21,7 +23,14 @@ const reducer = combineReducers({
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
 )
-const store = createStore(reducer, middleware)
+
+const store = createStore(reducer, localCart, middleware)
+
+store.subscribe(() => {
+  saveState({
+    localCart: store.getState().localCart
+  })
+})
 
 export default store
 export * from './user'
