+++ client/store/localCart.js
@@ -19,38 +19,6 @@ const gotCart = cart => ({
   cart
 })
 
-//THUNKS OF SORT... GET/set DATA FROM LOCAL STORAGE
-
-//load state from local storage
-export const loadState = () => {
-  return async dispatch => {
-    try {
-      const serializedState = localStorage.getItem('state')
-      if (serializedState === null) {
-        return undefined
-      }
-      console.log(serializedState)
-      //NEEDS TO CREATE ACTION CREATOR AND DISPATCH IT
-      dispatch(gotCart(JSON.parse(serializedState)))
-    } catch (err) {
-      return undefined
-    }
-  }
-}
-
-//puts state in local storage
-export const saveState = state => {
-  return async dispatch => {
-    try {
-      const serializedState = JSON.stringify(state)
-      localStorage.setItem('state', serializedState)
-      //NEEDS TO CREATE ACTION CREATOR AND DISPATCH IT
-    } catch {
-      // ignore write errors
-    }
-  }
-}
-
 const initialState = []
 
 //reducer
